+++ src/rgw/rgw_sal_sfs.cc
@@ -43,6 +43,7 @@
 #include "services/svc_zone_utils.h"
 
 #include "store/sfs/notification.h"
+#include "store/sfs/sfs_gc.h"
 #include "store/sfs/writer.h"
 
 #include "rgw/store/sfs/sqlite/sqlite_users.h"
@@ -382,11 +383,16 @@ int SFStore::initialize(
   const DoutPrefixProvider* dpp
 ) {
   ldpp_dout(dpp, 10) << __func__ << dendl;
+  gc = new sfs::SFSGC();
+  gc->initialize(cct, this);
+  gc->start_processor();
   return 0;
 }
 
 void SFStore::finalize(void) {
   ldout(ctx(), 10) << __func__ << ": TODO" << dendl;
+  delete gc;
+  gc = nullptr;
   return;
 }